refactor(snackbar/toast): optimize CSS
