+++ src/components/snackbar/themes/shared/snackbar.bootstrap.scss
@@ -3,13 +3,6 @@
 
 $theme: $bootstrap;
 
-[part~='base'] {
-    border-radius: var-get($theme, 'border-radius');
-    color: var-get($theme, 'text-color');
-    background: var-get($theme, 'background');
-    box-shadow: var-get($theme, 'elevation');
-}
-
 igc-button::part(base):hover {
     color: var-get($theme, 'button-color');
 }
