+++ src/components/toast/themes/shared/toast.bootstrap.scss
@@ -1,13 +1,5 @@
 @use 'styles/utilities' as *;
-@use '../light/themes' as *;
-
-$theme: $bootstrap;
 
 :host {
     padding: rem(11px);
-    color: var-get($theme, 'text-color');
-    background: var-get($theme, 'background');
-    border: #{rem(1px)} solid var-get($theme, 'border-color');
-    box-shadow: var-get($theme, 'elevation');
-    border-radius: var-get($theme, 'border-radius');
 }
