+++ src/components/toast/themes/shared/toast.fluent.scss
@@ -1,15 +1,7 @@
 @use 'styles/utilities' as *;
-@use '../light/themes' as *;
-
-$theme: $fluent;
 
 :host {
     @include type-style('caption');
 
     padding: rem(8px) rem(12px);
-    color: var-get($theme, 'text-color');
-    background: var-get($theme, 'background');
-    border: #{rem(1px)} solid var-get($theme, 'border-color');
-    box-shadow: var-get($theme, 'elevation');
-    border-radius: var-get($theme, 'border-radius');
 }
