--- src/components/toast/themes/shared/toast.indigo.scss
@@ -1,13 +1,0 @@
-@use 'styles/utilities' as *;
-@use '../light/themes' as *;
-
-$theme: $indigo;
-
-:host {
-    padding: rem(10px) rem(16px);
-    color: var-get($theme, 'text-color');
-    background: var-get($theme, 'background');
-    border: #{rem(1px)} solid var-get($theme, 'border-color');
-    box-shadow: var-get($theme, 'elevation');
-    border-radius: var-get($theme, 'border-radius');
-}
